Fixed installecr command to update airflow settings file

+++ install_handler.py
@@ -15,7 +15,7 @@ from tabulate import tabulate
 
 from config_handler import get_archive_config, get_lucidum_dir, get_jinja_templates_dir, \
     get_docker_compose_tmplt_file, get_ecr_images, get_images_from_ecr, get_local_images, \
-    get_service_image_mapping_config
+    get_docker_compose_service_image_mapping_config, get_airflow_service_image_mapping_config
 from docker_service import load_docker_images, pull_docker_image, copy_files_from_docker_container, \
     remove_docker_image, list_docker_images, get_docker_image, stop_docker_compose_service
 from exceptions import AppError
@@ -206,40 +206,89 @@ class DockerImagesUpdater:
 
 
 class DockerImagesUpdaterWithNewRestart(DockerImagesUpdater):
-    service_image_mapping = {
+    docker_compose_service_image_mapping = {
         "web": "mvp1_backend",
         "connector-api": "connector-api",
         "action-manager": "action-manager",
     }
 
-    def __init__(self, ecr_images, copy_default, restart, update_compose_file=False) -> None:
+    airflow_service_image_mapping = {
+        "merger": "python/ml",
+    }
+
+    def __init__(self, ecr_images, copy_default, restart, update_files=False) -> None:
         super().__init__(ecr_images, copy_default, restart)
-        self.update_compose_file = update_compose_file
+        self.update_files = update_files
 
     def update_docker_compose_file(self):
         lucidum_dir = get_lucidum_dir()
         docker_compose_file_path = os.path.join(lucidum_dir, "docker-compose.yml")
+        if not os.path.isfile(docker_compose_file_path):
+            logger.warning("'{}' file does not exist", docker_compose_file_path)
+            return
         with open(docker_compose_file_path) as f:
             data = yaml.full_load(f)
-        mapping = get_service_image_mapping_config() or self.service_image_mapping
+        mapping = get_docker_compose_service_image_mapping_config() or self.docker_compose_service_image_mapping
         for component in self._ecr_images:
             service = self._get_service_by_image(mapping, component["name"])
             if service is None:
                 continue
-            if service in data["services"]:
-                data["services"][service]["image"] = f"{component['name']}:{component['version']}"
+            if service not in data["services"]:
+                logger.warning("'{}' service not found in docker compose file", service)
+                continue
+            data["services"][service]["image"] = f"{component['name']}:{component['version']}"
         with open(docker_compose_file_path, "w") as f:
             yaml.dump(data, f)
 
+    def update_airflow_settings_file(self):
+        lucidum_dir = get_lucidum_dir()
+        airflow_settings_file_path = os.path.join(lucidum_dir, "airflow", "dags", "settings.yml")
+        if not os.path.isfile(airflow_settings_file_path):
+            logger.warning("'{}' file does not exist", airflow_settings_file_path)
+            return
+        with open(airflow_settings_file_path) as f:
+            data = yaml.full_load(f)
+        mapping = get_airflow_service_image_mapping_config() or self.airflow_service_image_mapping
+        for component in self._ecr_images:
+            service = self._get_service_by_image(mapping, component["name"])
+            if service is not None:
+                if service not in data["global"]:
+                    logger.warning("'{}' service not found in airflow settings file", service)
+                    continue
+                data["global"][service]["version"] = component["version"]
+            elif component["name"].startswith("connector"):
+                connector_type = component["name"].split("-", 1)[-1]
+                if "connectors" not in data["global"]:
+                    logger.warning("No connectors found in airflow settings file")
+                    continue
+                if connector_type not in data["global"]["connectors"]:
+                    logger.warning("'{}' service not found in airflow settings file", component["name"])
+                    continue
+                data["global"]["connectors"][connector_type]["version"] = component["version"]
+            else:
+                if component["name"] not in data["global"]:
+                    logger.warning("'{}' service not found in airflow settings file", component["name"])
+                    continue
+                data["global"][component["name"]]["version"] = component["version"]
+        with open(airflow_settings_file_path, "w") as f:
+            yaml.dump(data, f)
+
     def restart(self) -> None:
-        if self.update_compose_file or self._restart:
+        if self.update_files or self._restart:
             lucidum_dir = get_lucidum_dir()
-            mapping = get_service_image_mapping_config() or self.service_image_mapping
+            docker_compose_file_path = os.path.join(lucidum_dir, "docker-compose.yml")
+            docker_compose_data = None
+            if os.path.isfile(docker_compose_file_path):
+                with open(docker_compose_file_path) as f:
+                    docker_compose_data = yaml.full_load(f)
+            mapping = get_docker_compose_service_image_mapping_config() or self.docker_compose_service_image_mapping
             for component in self._ecr_images:
                 service = self._get_service_by_image(mapping, component["name"])
                 if service is None:
                     continue
-                stop_docker_compose_service(lucidum_dir, service)
+                if docker_compose_data is not None:
+                    if service in docker_compose_data["services"]:
+                        stop_docker_compose_service(lucidum_dir, service)
                 image_tag = f"{component['name']}:{component['version']}"
                 try:
                     remove_docker_image(image_tag, force=True)
@@ -248,8 +297,9 @@ class DockerImagesUpdaterWithNewRestart(DockerImagesUpdater):
                 image = pull_docker_image(component["image"])
                 image.tag(image_tag)
                 image.reload()
-            if self.update_compose_file:
+            if self.update_files:
                 self.update_docker_compose_file()
+                self.update_airflow_settings_file()
             run_docker_compose()
 
     def _get_service_by_image(self, mapping: dict, image_name: str) -> str:
@@ -300,8 +350,8 @@ def install_ecr(components, copy_default, restart, get_images=get_ecr_images):
     update_docker_images()
 
 @logger.catch(onerror=lambda _: sys.exit(1))
-def install_image_from_ecr(images, copy_default, restart, update_compose_file=False):
-    update_docker_images = DockerImagesUpdaterWithNewRestart(images, copy_default, restart, update_compose_file)
+def install_image_from_ecr(images, copy_default, restart, update_files=False):
+    update_docker_images = DockerImagesUpdaterWithNewRestart(images, copy_default, restart, update_files)
     update_docker_images()
 
 @logger.catch(onerror=lambda _: sys.exit(1))
